chore(utils): create analytics event classes util and update cms button to use new util
