+++ src/fields/analytics/types.ts
@@ -1,0 +1,12 @@
+export type AnalyticsEvent = {
+  hasAnalyticsEvent?: boolean | null | undefined;
+  eventName?: string | null | undefined;
+  properties?:
+    | {
+        propertyName: string;
+        propertyValue: string;
+        id?: string | null | undefined;
+      }[]
+    | null
+    | undefined;
+};
